# perf: mark dna::complement as inline