This speeds up computing the complement on a long string by ~2.3x. Not sure why rust doesn't inline such a trivial function on its own.

I didn't manage to compile your benchmarks, hence I can only point to the results from my [own project](https://github.com/kloetzl/libdna/blob/master/rust/benches/Brevcomp.rs).

> dna4_revcomp            time:   [688.18 µs 690.09 µs 692.27 µs]                         
> dnax_revcomp            time:   [3.0615 ms 3.0705 ms 3.0815 ms]                          
> 
> rust-bio                time:   [12.029 ms 12.105 ms 12.212 ms]                     
> rust-bio inlined        time:   [5.0742 ms 5.0901 ms 5.1087 ms]                              

There is still a small overhead of Rust over C. I think Rust repeatedly checks that `a as usize` is a valid index in `COMPLEMENT`. While that is usually not a problem, it is here, because the loop is so small and the extra instructions clog the pipeline.