@kloetzl 2022-12-24
> Can you also do the same for the complement function in src/alphabet/rna.rs?

Sure, will do.

> Benchmarks require using nightly - did you try that?

Didn't know that. Will give it a go.