@dcroote 2022-12-31
Thanks but it looks like you added `#[inline]` to `revcomp` instead of `complement` in `rna.rs`