@kloetzl 2023-01-02
> Thanks but it looks like you added `#[inline]` to `revcomp` instead of `complement` in `rna.rs`

Thanks for the thorough review. I blame my lingering cold.